requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
python-docx==1.1.2
openai==1.58.1
tenacity==9.0.0
